requirements.txt
paho-mqtt==1.6.1
spotipy==2.20.0
python-dotenv==0.21.0
